+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - ???
+
+### Changed
+
+- Moved CSV parsing to `middlewares/csv-payload`
+- Parsing errors will result in an exception being thrown
+- Strict types added
+
 ## [1.0.0] - 2018-01-25
 
 ### Added